refactor: split provider tests by concern
